+++ prowler/providers/scaleway/scaleway_provider.py
@@ -348,6 +348,7 @@ class ScalewayProvider(Provider):
             ScalewayCredentialsError,
             ScalewaySessionError,
             ScalewayAuthenticationError,
+            ScalewayIdentityError,
         ) as error:
             logger.error(
                 f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"